+++ secure/base.sh
@@ -1,6 +1,25 @@
 #!/bin/bash
 function deleteOrLockUnnecessaryUsersAndGroups() {
 	logTip $FUNCNAME
+	if [[ $INTERACTIVE == "Y" ]]; then
+		cat <<EOF
+userdel games
+passwd -l dbus
+passwd -l nobody
+passwd -l ftp
+passwd -l mail
+passwd -l shutdown
+passwd -l halt
+passwd -l operator
+passwd -l sync
+passwd -l adm
+passwd -l lp
+EOF
+		read -p "Do you want to carry out the above orders?[Y/N]: " wantDeleteUsersAndGroups
+	fi
+	if [[ $wantDeleteUsersAndGroups == "N" || $wantDeleteUsersAndGroups == "n" ]]; then
+		return
+	fi
 	userdel games
 	passwd -l dbus
 	passwd -l nobody
@@ -17,6 +36,19 @@ function deleteOrLockUnnecessaryUsersAndGroups() {
 
 function setPrivileges() {
 	logTip $FUNCNAME
+	if [[ $INTERACTIVE == "Y" ]]; then
+		cat <<EOF
+	chattr +i /etc/passwd
+	chattr +i /etc/shadow
+	chattr +i /etc/group
+	chattr +i /etc/gshadow
+	chmod -R 700 /etc/rc.d/init.d/*
+EOF
+		read -p "Do you want to carry out the above orders?[Y/N]: " wantSetPrivileges
+	fi
+	if [[ $wantSetPrivileges == "N" || $wantSetPrivileges == "n" ]]; then
+		return
+	fi
 	chattr +i /etc/passwd
 	chattr +i /etc/shadow
 	chattr +i /etc/group
@@ -25,34 +57,21 @@ function setPrivileges() {
 	logSuccess "Privileges has set."
 }
 
-function updatePort() {
+function updateSSHConfig() {
 	logTip $FUNCNAME
-	cp /etc/ssh/sshd_config /etc/ssh/sshd.bak
-	while true; do
-		read -p "Please enter ssh port number: " Port
-		if [ -n "$Port" ]; then
-			if ((Port <= 65535 && Port >= 1024)); then
-				break
-			else
-				echo "Ports can only be pure numbers within 1024-65535."
-			fi
-		else
-			Port="22"
-			break
-		fi
-	done
-
-	echo "Port $Port" >>/etc/ssh/sshd_config
-	if [ "$OS_VERSION" -eq 7 ]; then
-		firewall-cmd --zone=public --add-port=$Port/tcp --permanent
-	else
-		iptables -I INPUT -p tcp --dport $Port -j ACCEPT
+	if [[ $INTERACTIVE == "Y" ]]; then
+		cat <<EOF
+UseDNS no
+GSSAPIAuthentication no
+X11Forwarding no
+ChrootDirectory /home/%u
+AllowTcpForwarding no
+EOF
+		read -p "Do you want to configure the above content om /etc/ssh/sshd_config?[Y/N]: " wantUpdateSSHConfig
 	fi
-	logSuccess "Port has updated."
-}
-
-function updateSSH() {
-	logTip $FUNCNAME
+	if [[ $wantUpdateSSHConfig == "N" || $wantUpdateSSHConfig == "n" ]]; then
+		return
+	fi
 	sed -i 's/#UseDNS yes/UseDNS no/g' /etc/ssh/sshd_config
 	sed -i 's/GSSAPIAuthentication yes/GSSAPIAuthentication no/' /etc/ssh/sshd_config
 	sed -i 's/X11Forwarding yes/X11Forwarding no/' /etc/ssh/sshd_config
@@ -65,6 +84,12 @@ EOF
 
 function closeCtrlAltDel() {
 	logTip $FUNCNAME
+	if [[ $INTERACTIVE == "Y" ]]; then
+		read -p "Do you want to close CtrlAltDel?[Y/N]: " wantCloseCtrlAltDel
+	fi
+	if [[ $wantCloseCtrlAltDel == "N" || $wantCloseCtrlAltDel == "n" ]]; then
+		return
+	fi
 	sed -i "s/ca::ctrlaltdel:\/sbin\/shutdown -t3 -r now/#ca::ctrlaltdel:\/sbin\/shutdown -t3 -r now/" /etc/inittab
 	sed -i 's/^id:5:initdefault:/id:3:initdefault:/' /etc/inittab
 	logSuccess "CtrlAltDel has closed."
@@ -72,6 +97,12 @@ function closeCtrlAltDel() {
 
 function closeIpv6() {
 	logTip $FUNCNAME
+	if [[ $INTERACTIVE == "Y" ]]; then
+		read -p "Do you want to close Ipv6?[Y/N]: " wantCloseIpv6
+	fi
+	if [[ $wantCloseIpv6 == "N" || $wantCloseIpv6 == "n" ]]; then
+		return
+	fi
 	echo "net.ipv6.conf.all.disable_ipv6 = 1" >>/etc/sysctl.conf
 	echo "net.ipv6.conf.default.disable_ipv6 = 1" >>/etc/sysctl.conf
 	echo 1 >/proc/sys/net/ipv6/conf/all/disable_ipv6
@@ -81,6 +112,12 @@ function closeIpv6() {
 
 function closeSELinux() {
 	logTip $FUNCNAME
+	if [[ $INTERACTIVE == "Y" ]]; then
+		read -p "Do you want to close SELinux?[Y/N]: " wantCloseSELinux
+	fi
+	if [[ $wantCloseSELinux == "N" || $wantCloseSELinux == "n" ]]; then
+		return
+	fi
 	sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' /etc/selinux/config
 	setenforce 0
 	logSuccess "SELinux has closed."
@@ -89,8 +126,7 @@ function closeSELinux() {
 function main() {
 	deleteOrLockUnnecessaryUsersAndGroups
 	setPrivileges
-	updatePort
-	updateSSH
+	updateSSHConfig
 	closeCtrlAltDel
 	closeIpv6
 	closeSELinux
